test(integer): share image path test helpers
Follow-up to #470 review feedback. Move findPath/pathHasEntry out of airflow_image_test.go into helpers_test.go so rabbitmq_cluster_operator_image_test.go does not rely on helper definitions hidden in an unrelated Airflow test file.
